+++ src/level2/procedures/Parametrized.java
@@ -3,5 +3,5 @@ package level2.procedures;
 import level2.constants.Executable;
 
 public interface Parametrized extends Executable {
-    Executable getFunction(byte... parameter);
+    Executable getFunction(short... parameter);
 }
